patch

+++ api/chatgpt.py
@@ -10,8 +10,8 @@ class ChatGPT:
     def __init__(self):
         self.prompt = Prompt()
         self.model = os.getenv("OPENAI_MODEL", default="text-davinci-003")
-        self.temperature = int(os.getenv("OPENAI_TEMPERATURE", default=0))
-        self.frequency_penalty = int(os.getenv("OPENAI_FREQUENCY_PENALTY", default=0))
+        self.temperature = float(os.getenv("OPENAI_TEMPERATURE", default=0))
+        self.frequency_penalty = float(os.getenv("OPENAI_FREQUENCY_PENALTY", default=0))
         self.presence_penalty = float(os.getenv("OPENAI_PRESENCE_PENALTY", default=0.6))
         self.max_tokens = int(os.getenv("OPENAI_MAX_TOKENS", default=240))
